requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
fastapi==0.119.1
fonttools==4.59.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
imageio==2.37.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.6
multidict==6.7.0
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.12.0.88
packaging==25.0
pillow==11.3.0
playwright==1.55.0
postgrest==2.22.1
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.22.1
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
storage3==2.22.1
StrEnum==0.4.15
supabase==2.22.1
supabase-auth==2.22.1
supabase-functions==2.22.1
threadpoolctl==3.6.0
tifffile==2025.8.28
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
